api is posting and returning the url

+++ client/src/App.js
@@ -4,7 +4,7 @@ import Login from './Login';
 import HomeContainer from './HomeContainer';
 import Account from './Account';
 import Navbar from './Navbar';
-import Cart from './Cart';
+// import Cart from './Cart';
 
 import './App.css';
 import CartContainer from './CartContainer';
@@ -15,6 +15,8 @@ function App() {
   const [user, setUser] = useState([])
   const [fullPartList, setFullPartList] = useState([])
   const [cartArr, setCartArr] = useState([])
+  const [editPartForm, setEditPartForm] = useState('')
+  // console.log(editPartForm)
 
   
   useEffect(() => {
@@ -54,10 +56,10 @@ function App() {
       fetch("/user_carts", config)
         .then((r) => r.json())
         .then((data) => setCartArr([...cartArr, data]));
-        console.log(cartArr)
+        // console.log(cartArr)
     }
 
-    function handleAddPart(e, user, title, price, quantity, year, make, model,  image, description) {
+    function handleAddPart(e, user, title, price, quantity, year, make, model,  imageURL, description) {
       e.preventDefault()
       console.log('clicked')
       const partItem = {
@@ -68,7 +70,7 @@ function App() {
         year: year,
         price: price,
         quantity: quantity,
-        image: image,
+        image: imageURL,
         description: description
       };
       let config = {
@@ -84,6 +86,7 @@ function App() {
         .then((data) => setFullPartList([...fullPartList, data]));
         console.log(fullPartList)
         console.log("added")
+
     }
 
 
@@ -96,26 +99,66 @@ function App() {
   }, []);
 
   function handleLogoutClick(){
-    fetch('/logout', {
+    if (window.confirm("Are you sure you want to Logout?")) {
+     fetch('/logout', {
       method: 'DELETE'
     })
+    
     .then(res => {
       if (res.ok) {
         setUser()
         history.push('/')
       }
     })
+    }
   }
 
   function handleDeleteClick(id) {
-    fetch(`/user_cart/${id}`, {
+    if (window.confirm("Are you sure you want to Remove from your Watchlist?")) {
+    fetch(`/user_carts/${id}`, {
       method: "DELETE",
     });
     const filteredCart = cartArr.filter((item) => item.id !== id);
     setCartArr(filteredCart);
   }
+  }
 
-  
+  function removeListing(id) {
+    if (window.confirm("Are you sure you want to Remove this Listing?")) {
+    fetch(`/parts/${id}`, {
+      method: "DELETE",
+    });
+    const filteredList = fullPartList.filter((item) => item.id !== id);
+    setFullPartList(filteredList);
+    }
+  }
+  function editListing (e, list, user, editForm){
+    e.preventDefault()
+    const partItem = {
+      user_id: user.id,
+      title: editForm.title,
+      make: editForm.make,
+      model: editForm.model,
+      year: editForm.year,
+      price: editForm.price,
+      quantity: editForm.quantity,
+      image: editForm.image,
+      description: editForm.description
+    };
+    let config = {
+      method: "PATCH",
+      headers: {
+        Accept: "application/json",
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify(partItem),
+    };
+    fetch(`/parts/${list.id}`, config)
+      .then((r) => r.json())
+      .then((data) => setFullPartList([...fullPartList, data]));
+      
+   
+  }
 
   const onLogin = (user) => {
     setUser(user)
@@ -127,7 +170,7 @@ function App() {
 
   return (
     <div className="App">
-      <Navbar handleLogoutClick={handleLogoutClick}/>
+      <Navbar handleLogoutClick={handleLogoutClick} user={user}/>
       <Switch>
       <Route exact path="/login">
        <Login setUser={setUser} onLogin={onLogin}  />
@@ -136,9 +179,9 @@ function App() {
         <HomeContainer fullPartList={fullPartList} handleAddToCart={handleAddToCart} user={user}/>
        </Route> 
        <Route exact path="/account">
-        <Account user={user}  handleAddPart={handleAddPart}/>
+        <Account user={user} setEditPartForm={setEditPartForm} editPartForm={editPartForm} handleAddPart={handleAddPart} full={fullPartList} removeListing={removeListing} editListing={editListing}/>
        </Route> 
-       <Route exact path="/cart">
+       <Route exact path="/watchlist">
         <CartContainer user={user} cartArr={cartArr} onDelete={handleDeleteClick}/>
        </Route>
